Feat: 북마크 관련 에러 핸들링
